% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexes.R
\name{CVG}
\alias{CVG}
\title{Coefficient of Velocity of Germination}
\usage{
CVG(eval_days)
}
\arguments{
\item{eval_days}{Germination evalaluation days}
}
\value{
Coefficient of Velocity of Germination
}
\description{
This function Calculate the Coefficient of Velocity of Germination
}

