% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ger_factors.R
\name{evalDays}
\alias{evalDays}
\title{Select Evaluation Days}
\usage{
evalDays(evalName, data)
}
\arguments{
\item{evalName}{Prefix of the evaluation variable}

\item{data}{Data with germination values}
}
\value{
Evaluation columns
}
\description{
Give matrix with the evaluation days
}
\examples{
\dontrun{ 
library(GerminaR)
dt <- prosopis
dm <- evalDays(evalName = "D", data = dt)
dm
}
}
