% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{evalFact}
\alias{evalFact}
\title{Select Factors of germination matrix}
\usage{
evalFact(data)
}
\arguments{
\item{data}{Data with germination values}
}
\value{
Factor colums
}
\description{
Give matrix with the factor
}

