% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexes.R
\name{ger_ASG}
\alias{ger_ASG}
\title{ArcSin of Germination Percentage}
\usage{
ger_ASG(SeedN, evalDays)
}
\arguments{
\item{SeedN}{Name of the colon with the seed numbers}

\item{evalDays}{Germination evalaluation days}
}
\value{
ArcSin of Germination
}
\description{
This function calculate the arcsin of germination percentage for normalization
}

