% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexes.R
\name{ger_CVL}
\alias{ger_CVL}
\title{Coefficient of Velocity of Germination}
\usage{
ger_CVL(evalDays)
}
\arguments{
\item{evalDays}{Germination evalaluation days}
}
\value{
Coefficient of Velocity of Germination
}
\description{
This function calculate the coefficient of velocity of germination according at the time lapse of the evaluations
}

