% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{ger_day}
\alias{ger_day}
\title{Cumulative sum of germination by period of time for line graphic}
\usage{
ger_day(data, evalName)
}
\arguments{
\item{data}{Data with the germination avaliation process}
}
\value{
Data frame with the germination by period
}
\description{
This function made a data table with the cumulative sum of values of germination by days.
}

