% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexes.R
\name{ger_GRP}
\alias{ger_GRP}
\title{Germination Seed percentage}
\usage{
ger_GRP(seeds, eval_days)
}
\arguments{
\item{seeds}{Total seed sown}

\item{eval_days}{Germination evaluation days}
}
\value{
Percentage of seed germinated
}
\description{
This function Calculate the Germination percentage
}

