% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexes.R
\name{ger_GRS}
\alias{ger_GRS}
\title{Germinated Seed Number}
\usage{
ger_GRS(evalDays)
}
\arguments{
\item{evalDays}{Germination evaluation days}
}
\value{
Number of seed germianated
}
\description{
This function calculate the number of seed germinated.
}

