% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexes.R
\name{ger_GRU}
\alias{ger_GRU}
\title{Germination Uncertainty}
\usage{
ger_GRU(eval_days)
}
\arguments{
\item{eval_days}{Germination evalaluation days}
}
\value{
Germination Uncertainty
}
\description{
This function calculate the germination uncertainty of the germination process.
}

