% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexes.R
\name{ger_GSI}
\alias{ger_GSI}
\title{Germination Synchronization Index}
\usage{
ger_GSI(evalDays)
}
\arguments{
\item{evalDays}{Germination evalaluation days}
}
\value{
Germination Synchrony
}
\description{
This function calculate the germination synchronization of the germination process
}

