% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ger_indices.R
\name{ger_GSP}
\alias{ger_GSP}
\title{Germination Speed}
\usage{
ger_GSP(evalName, data)
}
\arguments{
\item{evalName}{Prefix of the names of the periods of evaluation.}

\item{data}{The name of the data frame containing the data.}
}
\value{
It returns an vector with the Germination Speed
}
\description{
This function calculates the Germination Speed according at the time lapse of the evaluations.
}
\examples{
library(GerminaR)
dfr <- prosopis
gsp <- ger_GSP(evalName = "D", data = dfr)
gsp
}
