% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{ger_intime}
\alias{ger_intime}
\title{Cumulative sum of germination by period of time for line graphic}
\usage{
ger_intime(SeedN, evalName, data)
}
\arguments{
\item{SeedN}{Name of the colonn with the seed numbers}

\item{evalName}{Prefix of the evalaution variable}

\item{data}{Data with the germination avaliation process}
}
\value{
Data frame with the germination by period
}
\description{
This function made a data table with the cumulative sum of values of germination by days.
}
\details{
Need a summary by factor before use it with function SummaryBy.
}

