% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexes.R
\name{ger_MGR}
\alias{ger_MGR}
\title{Mean Germination Rate}
\usage{
ger_MGR(eval_days)
}
\arguments{
\item{eval_days}{Germination evalaluation days}
}
\value{
Mean Germination Rate
}
\description{
This function Calculate the Mean Germination Rate
}

