% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexes.R
\name{ger_MGT}
\alias{ger_MGT}
\title{Mean Germination Time}
\usage{
ger_MGT(evalDays)
}
\arguments{
\item{evalDays}{Germination evalaluation days}
}
\value{
Mean Germination Time
}
\description{
This function calculate the mean germination time of germination according at the time lapse of the evaluations
}

