% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexes.R
\name{ger_SDG}
\alias{ger_SDG}
\title{Standard deviation of the Germination Time}
\usage{
ger_SDG(eval_days)
}
\arguments{
\item{eval_days}{Germination evalaluation days}
}
\value{
Standard deviation of the Germination Time
}
\description{
This function Calculate the Standard deviation of the Germination Time
}

