% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexes.R
\name{ger_VGT}
\alias{ger_VGT}
\title{Variance of the Mean Germination Time}
\usage{
ger_VGT(evalDays)
}
\arguments{
\item{evalDays}{Germination evalaluation days}
}
\value{
Variance of Germination
}
\description{
This function calculate the variance of the mean germination time.
}

