% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{germcalculate}
\alias{germcalculate}
\title{Calculation of the variables data matrix}
\usage{
germcalculate(fieldbook)
}
\arguments{
\item{fieldbook}{Data frame with the germination evaluation}
}
\value{
Names of the evaluation days
}
\description{
This function Calculate the Number of Germinated Seeds
}
\author{
Omar Benites
}

