% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GerminaQuant.R
\name{GerminaQuant}
\alias{GerminaQuant}
\title{GerminaQuant for R}
\source{
Lozano-Isla, F., Benites Alfaro, O., & Pompelli, M. F. (2016).
GerminaQuant for R (Patent No. BR 51 2016 001327-3).
https://inkaverse.shinyapps.io/germinaquant/
}
\usage{
GerminaQuant(dependencies = FALSE)
}
\arguments{
\item{dependencies}{Install package dependencies for run the app}
}
\description{
GermiQuant for R app allows make the calculation for the germination indices
incredibly easy in a interactive applications build in base a GerminaR R
package and Shiny. GermiQuant app is live!. Outputs change instantly as users
modify inputs, without requiring a reload the app. The principal features of
the application allow calculate the principal germination indices,
statistical analysis and easy way to plot the results.
}
\concept{GerminaR}
