% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germ_datasets.R
\docType{data}
\name{GerminaR}
\alias{GerminaR}
\title{Germination Data}
\format{A data frame with 30 columns and 66 rows.}
\source{
UFRPE experimental data.
}
\usage{
GerminaR
}
\description{
Dataset containing information from a experiremnt with 6 genotypes Jatropha curcas under  3 levels of salinity stress evaluated during 25 days.
}
\references{
This data set was developed in UFRPE (2015)
}

