% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_eval_days}
\alias{get_eval_days}
\title{Evaluation days row names}
\usage{
get_eval_days(fieldbook)
}
\arguments{
\item{fieldbook}{Data frame with the germination evaluation}
}
\value{
Names of the evaluation days
}
\description{
This function Calculate the Number of Germinated Seeds
}
\author{
Omar Benites
}

