% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexes.R
\name{GRS}
\alias{GRS}
\title{Germination Seed Number}
\usage{
GRS(eval_days)
}
\arguments{
\item{eval_days}{Germination evaluation days}
}
\value{
Number of seed germinated
}
\description{
This function Calculate the Number of Germinated Seeds
}
\author{
Flavio Lozano
}

