% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexes.R
\name{GRU}
\alias{GRU}
\title{Germination Uncertainty}
\usage{
GRU(eval_days)
}
\arguments{
\item{eval_days}{Germination evalaluation days}
}
\value{
Germination Uncertainty (U)
}
\description{
This function Calculate the Germination Uncertainty
}

