% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexes.R
\name{GSI}
\alias{GSI}
\title{Germination Synchronization Index}
\usage{
GSI(eval_days)
}
\arguments{
\item{eval_days}{Germination eval_days}
}
\value{
Germination Synchronization Index (Z)
}
\description{
This function Calculate the Germination Synchronization Index
}
\author{
Flavio Lozano
}

