% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H2cal.R
\name{H2cal}
\alias{H2cal}
\title{Heritability in plant breeding on a genotype-difference basis}
\usage{
H2cal(
  data,
  trait,
  gen.name,
  rep.n,
  loc.name = NULL,
  year.name = NULL,
  loc.n = 1,
  year.n = 1,
  fix.model,
  ran.model,
  blues = TRUE
)
}
\arguments{
\item{data}{Experimental design data frame with the factors and traits.}

\item{trait}{Name of the trait.}

\item{gen.name}{Name of the genotypes.}

\item{rep.n}{Number of replications in the experiment.}

\item{loc.name}{Name of the location (default = NULL). See details.}

\item{year.name}{Name of the years (default = NULL). See details.}

\item{loc.n}{Number of locations (default = 1). See details.}

\item{year.n}{Number of years (default = 1). See details.}

\item{fix.model}{The fixed effects in the model. See examples.}

\item{ran.model}{The random effects in the model. See examples.}

\item{blues}{Calculate the BLUEs (default = TRUE).}
}
\value{
A list with three object: 
1. Table with the variance components and hertitabilities
2. BLUPs
3. BLUEs
}
\description{
Function to calculate:
1. The variance components.
2. Heritability under three approachs: Standart, Cullis and Piepho.
3. Best Linear Unbiased Predictors (BLUPs).
4. Best Linear Unbiased Estimators (BLUEs).
}
\details{
The function allows to made the calculation for individual or multi-enviomental trials (MET) using th fixed and random model.
For individual experiments is necesary provide the \code{trait}, \code{gen.name}, \code{rep.n}.
For MET experiments you should \code{loc.n} and \code{loc.name} y/o \code{year.n} and \code{year.name} according your experiment. For more information review the references.
The blues calculation is based in the pairwaise comparison and its could takes time according the number of the genotypes. You can especify as \code{blues = FALSE} for calculate the variance componets and blups faster.
}
\examples{
library(agridat)
hr <- H2cal(data = john.alpha
            , blues = F
            , trait = "yield"
            , gen.name = "gen"
            , rep.n = 3
            , fix.model = "rep + (1|rep:block) + gen"
            , ran.model = "rep + (1|rep:block) + (1|gen)"
            )
hr$tabsmr 
}
\references{
Schmidt, P., J. Hartung, J. Bennewitz, and H.-P. Piepho. 2019. Heritability in Plant Breeding on a Genotype-Difference Basis. Genetics 212(4): 991–1008. doi: 10.1534/genetics.119.302134.

Schmidt, P., J. Hartung, J. Rath, and H.-P. Piepho. 2019. Estimating Broad-Sense Heritability with Unbalanced Data from Agricultural Cultivar Trials. Crop Science 59(2): 525–536. doi: 10.2135/cropsci2018.06.0376.
}
\author{
Belen Kistner

Flavio Lozano-Isla
}
