% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{metamorphosis}
\alias{metamorphosis}
\title{Transfor dataframe based in a dictionary}
\usage{
metamorphosis(fielbook, dictionary, from, to, index, colnames)
}
\arguments{
\item{dictionary}{data frame with at least 3 colums (original names, new names and variable type)}

\item{from}{Column name of a data frame with the original names of the variables}

\item{to}{Column name of a data frame with the new names of the variables}

\item{index}{Column name of a data frame with the type and level from the variables}

\item{colnames}{Character vector with the column names}

\item{fieldbook}{data frame with the original information}
}
\value{
List with original fieldbook, variables and new fieldbook
}
\description{
Transfor all fieldbook data frame according to data dictionary
}
