% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexes.R
\name{MGR}
\alias{MGR}
\title{Mean Germination Rate}
\usage{
MGR(values)
}
\arguments{
\item{values}{Germination values}
}
\value{
Mean Germination Rate
}
\description{
This function Calculate the Mean Germination Rate
}
\author{
Flavio Lozano
}

