% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexes.R
\name{MGT}
\alias{MGT}
\title{Mean Germination Time}
\usage{
MGT(values)
}
\arguments{
\item{values}{Germination values}
}
\value{
ArcSin of Germination Values
}
\description{
This function Calculate the ArcSin of Germination Values for normalization
}
\author{
Flavio Lozano
}

