% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmp.R
\name{osmp}
\alias{osmp}
\title{Osmotic potential calculator}
\usage{
osmp(type = "salt", vol, pres, temp, mw, ki)
}
\arguments{
\item{type}{Salt or PEG-6000 c("salt", "peg6000"). Default: "salt".}

\item{vol}{volume (liters)}

\item{pres}{Pressure (Mpa) in negative values. 1 bar = 0.1 Mpa}

\item{temp}{Temperature (centigrade)}

\item{mw}{Molecular weight}

\item{ki}{Salt dissociation constant (NaCl = 1.8)}
}
\value{
Numeric value (grams)
}
\description{
Function to calculate the grams of salt or PEG-6000 needed for determinated osmotic potential
}
