% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot}
\alias{plot}
\title{Plot line or bar graphic}
\usage{
plot(data, type = c("bar", "line"), x, y, z, lmt, brk, lgd = c(0.93, 0.77),
  ylab = "", xlab = "", lgl = "", sig = sg)
}
\arguments{
\item{data}{Output dtsm fuction}

\item{type}{Type of graphic. "bar" or "line"}

\item{x}{Axis x variable}

\item{y}{Axis t variable}

\item{z}{variable for color and shape}

\item{lmt}{Limit of the axis y for the graphic. ie c(0,100)}

\item{brk}{Brake in the axis y for the graphic. ie 0:100*20}

\item{lgd}{Possition of the legend. Default c(0.93,0.77). If "none" will not have legend}

\item{ylab}{Title for the axis y}

\item{xlab}{Title for the axis x}

\item{lgl}{Title for the legend}

\item{sig}{Significance of the result (letters)}
}
\value{
Line o bar plot
}
\description{
Function use the dtsm funtion for plot the results
}

