% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rep_row}
\alias{rep_row}
\title{Repeated Rows in a data matrix}
\usage{
rep_row(Rseq, Nrow)
}
\arguments{
\item{Rseq}{Row sequence for the data matrix}

\item{Nrow}{Number of rows for the data matrix}
}
\value{
Data Matrix with day of the germination
}
\description{
This function made a data table with the evaluation days of germination
}
