% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{runGerminaR}
\alias{runGerminaR}
\title{@description When the function is invoked, GerminaR display a shiny app application showing 
 web tabs about import, check, analyze and export germination data.}
\usage{
runGerminaR()
}
\description{
@description When the function is invoked, GerminaR display a shiny app application showing 
 web tabs about import, check, analyze and export germination data.
}

