% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sbsum}
\alias{sbsum}
\title{This function allows sum columns and avoid NA values at the same time}
\usage{
sbsum(x)
}
\arguments{
\item{x}{vector of measurements}
}
\value{
A vector contains sum of values that avoid NA's
}
\description{
This function allows sum columns and avoid NA values at the same time
}
\author{
Omar Benites
}
\keyword{NA's}
\keyword{sum,}

