% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{starts_with}
\alias{starts_with}
\title{Select colum according the initial letters of the words}
\usage{
starts_with(vars, match, ignore.case = TRUE)
}
\arguments{
\item{vars}{vars}

\item{match}{match}

\item{ignore.case}{case}
}
\value{
Matrix with the select colum
}
\description{
Select colum according the initial letters of the words
}
\details{
https://github.com/hadley/dplyr/blob/50309db8f04cbcc87e4568a4bfa1f0c718e824c9/R/select-utils.R
}
\author{
Hadley Wickham
}

