% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{Summary_Ger}
\alias{Summary_Ger}
\title{Summary of germination variables}
\usage{
Summary_Ger(data)
}
\arguments{
\item{data}{Data with the germination avaliation}
}
\value{
Data frame with the values of germination variables
}
\description{
This function made a data table with the values of germination variables
}

