% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{tb_web}
\alias{tb_web}
\title{HTML tables for markdown documents}
\usage{
tb_web(data, title = NULL, digits = 3, rnames = FALSE)
}
\arguments{
\item{data}{dataset}

\item{title}{Title for the table}

\item{digits}{digits number in the table exported}

\item{rnames}{row names}
}
\value{
table in markdown format for html documents
}
\description{
Export tables with download, pasta and copy buttons
}
