% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexes.R
\name{VGT}
\alias{VGT}
\title{Variance of Germination Time}
\usage{
VGT(eval_days)
}
\arguments{
\item{eval_days}{Germination evalaluation days}
}
\value{
Variance of Germination Time
}
\description{
This function Calculate the Variance of Germination Time
}

