% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexes.R
\name{ASG}
\alias{ASG}
\title{ArcSin of Germination Values}
\usage{
ASG(seeds, values)
}
\arguments{
\item{seeds}{Total seed sown}

\item{values}{Germination values}
}
\value{
ArcSin of Germination Values
}
\description{
This function Calculate the ArcSin of Germination Values for normalization
}
\author{
Flavio Lozano
}

