% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexes.R
\name{ASG}
\alias{ASG}
\title{ArcSin of Germination eval_days}
\usage{
ASG(seeds, eval_days)
}
\arguments{
\item{seeds}{Total seed sown}

\item{eval_days}{Germination eval_days}
}
\value{
ArcSin of Germination eval_days
}
\description{
This function Calculate the ArcSin of Germination eval_days for normalization
}
\author{
Flavio Lozano
}

