% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexes.R
\name{CoVG}
\alias{CoVG}
\title{Coefficient of Variation of the Mean Germination Time}
\usage{
CoVG(values)
}
\arguments{
\item{values}{Germination values}
}
\value{
Coefficient of Variation of the Mean Germination Time
}
\description{
This function Calculate the Coefficient of Variation of the Mean Germination Time
}
\author{
Flavio Lozano
}

