% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dtsm}
\alias{dtsm}
\title{Mean Comparison Table Summary}
\usage{
dtsm(meanComp)
}
\arguments{
\item{meanComp}{Object list with the result from mean comparison test}
}
\value{
Table with complete data for graphics
}
\description{
Function using resulting output from mean comparison test from agricolae package optimized for graphs.
}

