% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{evalDays}
\alias{evalDays}
\title{Select Evaluation Days}
\usage{
evalDays(data, evalName)
}
\arguments{
\item{data}{Data with germination values}

\item{evalName}{Prefix of the evalaution variable}
}
\value{
Evalution colums
}
\description{
Give matrix with the evaluation days
}

