% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{evalDays}
\alias{evalDays}
\title{Select Evaluation Days}
\usage{
evalDays(evalName, data)
}
\arguments{
\item{evalName}{Prefix of the evalaution variable}

\item{data}{Data with germination values}
}
\value{
Evalution colums
}
\description{
Give matrix with the evaluation days
}
\examples{
\dontrun{ 
library(GerminaR)
dt <- GerminaR
dm <- evalDays(evalName = "Ev", data = dt)
dm
}
}

