% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{evalFactor}
\alias{evalFactor}
\title{Select Factors of germination matrix}
\usage{
evalFactor(evalName, data)
}
\arguments{
\item{evalName}{Prefix of the evalaution variable}

\item{data}{Data with germination values}
}
\value{
Factor columsl
}
\description{
Give matrix with the factor
}
\examples{
\dontrun{ 
library(GerminaR)
dt <- prosopis
dm <- evalFactor(evalName = "D", data = dt)
dm
}
}
