% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{evalFactor}
\alias{evalFactor}
\title{Select Factors of germination matrix}
\usage{
evalFactor(data, evalName)
}
\arguments{
\item{data}{Data with germination values}

\item{evalName}{Prefix of the evalaution variable}
}
\value{
Factor colums
}
\description{
Give matrix with the factor
}

