% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexes.R
\name{ger_ASG}
\alias{ger_ASG}
\title{ArcSin of Germination eval_days}
\usage{
ger_ASG(seeds, eval_days)
}
\arguments{
\item{seeds}{Total seed sown}

\item{eval_days}{Germination evalaluation days}
}
\value{
ArcSin of Germination percentage
}
\description{
This function Calculate the ArcSin of Germination eval_days for normalization
}

