% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ger_indices.R
\name{ger_ASG}
\alias{ger_ASG}
\title{ArcSin of Germination Percentage}
\usage{
ger_ASG(SeedN, evalName, data)
}
\arguments{
\item{SeedN}{Name of the column with the number of seeds sown.}

\item{evalName}{Prefix of the names of the periods of evaluation.}

\item{data}{The name of the data frame containing the data.}
}
\value{
It returns an vector with the ArcSin of Germination values
}
\description{
This function calculates the arcsin of germination percentage for normalization.
}
\examples{
library(GerminaR)
dfr <- prosopis
gas <- ger_ASG(SeedN = "seeds", evalName = "D", data = dfr)
gas
}
