% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{ger_cumsum}
\alias{ger_cumsum}
\title{Cumulative sum of germination matrix}
\usage{
ger_cumsum(SeedN, evalName, method = c("percentage", "relative"), data)
}
\arguments{
\item{SeedN}{Name of the colonn with the seed numbers}

\item{evalName}{Prefix of the evalaution variable}

\item{method}{Type of cummulative germination}

\item{data}{Data with the germination avaliation process}
}
\value{
Data frame with the cumulative sum
}
\description{
This function made a data table with the cumulative sum of values of germination.
}

