% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexes.R
\name{ger_CVG}
\alias{ger_CVG}
\title{Coefficient of Variation of the Mean Germination Time}
\usage{
ger_CVG(eval_days)
}
\arguments{
\item{eval_days}{Germination evalaluation days}
}
\value{
Coefficient of Variation of the Mean Germination Time
}
\description{
This function Calculate the Coefficient of Variation of the Mean Germination Time
}

