% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexes.R
\name{ger_CVG}
\alias{ger_CVG}
\title{Coefficient of Variance of the Mean Germination Time}
\usage{
ger_CVG(eval_days)
}
\arguments{
\item{eval_days}{Germination evalaluation days}
}
\value{
Coefficient of Variance of germination
}
\description{
This function calculate the coefficient of variation of the mean germination time
}

