% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ger_indices.R
\name{ger_CVG}
\alias{ger_CVG}
\title{Coefficient of Variance of the Mean Germination Time}
\usage{
ger_CVG(evalName, data)
}
\arguments{
\item{evalName}{Prefix of the names of the periods of evaluation.}

\item{data}{The name of the data frame containing the data.}
}
\value{
It returns an vector with the values of Coefficient of Variance of germination
}
\description{
This function calculates the coefficient of variation of the mean germination time
}
\examples{
library(GerminaR)
dfr <- prosopis
cvg <- ger_CVG(evalName = "D", data = dfr)
cvg
}
