% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexes.R
\name{ger_CVL}
\alias{ger_CVL}
\title{Coefficient of Velocity of Germination}
\usage{
ger_CVL(evalName, freq = 1, data)
}
\arguments{
\item{evalName}{Prefix of the names of the periods of evaluation.}

\item{freq}{Frecuency of evualtion of the experimental units (days or hour)}

\item{data}{The name of the data frame containing the data.}
}
\value{
It returns an vector with the Coefficient of Velocity of Germination
}
\description{
This function calculate the coefficient of velocity of germination according at the time lapse of the evaluations.
}

