% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexes.R
\name{ger_GRS}
\alias{ger_GRS}
\title{Germinated Seed Number}
\usage{
ger_GRS(eval_days)
}
\arguments{
\item{eval_days}{Germination evaluation days}

\item{seeds}{Total seed sown}
}
\value{
Number of seed germianated in the experimen for EU
}
\description{
This function Calculate the Germination percentage
}

