% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ger_indices.R
\name{ger_GRS}
\alias{ger_GRS}
\title{Germinated Seed Number}
\usage{
ger_GRS(evalName, data)
}
\arguments{
\item{evalName}{Prefix of the names of the periods of evaluation.}

\item{data}{The name of the data frame containing the data.}
}
\value{
Number of seed germinated
}
\description{
This function calculates the number of seed germinated.
}
\examples{

library(GerminaR)
dt <- prosopis
grs <- ger_GRS(evalName = "D", data = dt)
grs
}
