% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexes.R
\name{ger_GSI}
\alias{ger_GSI}
\title{Germination Synchronization Index}
\usage{
ger_GSI(eval_days)
}
\arguments{
\item{eval_days}{Germination evalaluation days}
}
\value{
Germination Synchronization Index (Z)
}
\description{
This function Calculate the Germination Synchronization Index
}

