% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{ger_intime}
\alias{ger_intime}
\title{Cumulative sum of germination by period of time for line graphic}
\usage{
ger_intime(Factor, SeedN, evalName, method = c("percentage", "relative"),
  data)
}
\arguments{
\item{Factor}{Factor wich will be graph in time}

\item{SeedN}{Name of the colonn with the seed numbers}

\item{evalName}{Prefix of the evalaution variable}

\item{method}{Type of cummulative germination}

\item{data}{Data with the germination avaliation process}
}
\value{
Data frame with the germination by period
}
\description{
This function made a data table with the cumulative sum of values of germination by days.
}
\details{
Need a summary by factor before use it with function SummaryBy.
}
\examples{

library(GerminaR)
library(ggplot2)

dt <- GerminaR
gnt <- ger_intime(Factor= "Genotype", SeedN = "NSeeds", evalName = "Ev", method = "percentage", data = dt)

ggplot(gnt, aes(variable , value, colour = Genotype, group = Genotype)) +
  geom_line() +
  geom_point(size=1)+
  theme_bw()+
  ylab("Germination (\%)")+
  xlab("days")
}

