% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{ger_intime}
\alias{ger_intime}
\title{Cumulative sum of germination by period of time for line graphic}
\usage{
ger_intime(Factor, SeedN, evalName, method = "percentage", data)
}
\arguments{
\item{Factor}{Factor wich will be graph in time}

\item{SeedN}{Name of the colonn with the seed numbers}

\item{evalName}{Prefix of the evalaution variable}

\item{method}{Type of cummulative germination. "percentage" or "relative"}

\item{data}{Data with the germination avaliation process}
}
\value{
Data frame with the germination by period
}
\description{
This function makes a data table with the cumulative sum of values of germination by days.
}
\details{
Need a summary by factor before use it with function SummaryBy.
}
\examples{

library(GerminaR)
dt <- prosopis
grt <- ger_intime(Factor = "nacl", SeedN = "seeds", 
                  evalName = "D", method = "percentage", data = dt)
head(grt, 10)
}
