% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexes.R
\name{ger_SDG}
\alias{ger_SDG}
\title{Standard deviation of the Mean Germination Time}
\usage{
ger_SDG(eval_days)
}
\arguments{
\item{eval_days}{Germination evalaluation days}
}
\value{
Standard desviation of germination
}
\description{
This function calculate the standard desviation of the mean germination time
}

