% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexes.R
\name{ger_SDG}
\alias{ger_SDG}
\title{Standard deviation of the Mean Germination Time}
\usage{
ger_SDG(evalName, freq = 1, data)
}
\arguments{
\item{evalName}{Prefix of the names of the periods of evaluation.}

\item{freq}{Frecuency of evualtion of the experimental units (days or hour)}

\item{data}{The name of the data frame containing the data.}
}
\value{
It returns an vector with the values of Standard desviation of germination
}
\description{
This function calculate the standard desviation of the mean germination time
}

