% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ger_indices.R
\name{ger_SDG}
\alias{ger_SDG}
\title{Standard deviation of the Mean Germination Time}
\usage{
ger_SDG(evalName, data)
}
\arguments{
\item{evalName}{Prefix of the names of the periods of evaluation.}

\item{data}{The name of the data frame containing the data.}
}
\value{
It returns an vector with the values of Standard deviation of germination
}
\description{
This function calculates the standard deviation of the mean germination time
}
\examples{
library(GerminaR)
dfr <- prosopis
sdg <- ger_SDG(evalName = "D", data = dfr)
sdg
}
