% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{ger_summary}
\alias{ger_summary}
\title{Summary of Germination Variables}
\usage{
ger_summary(SeedN, evalName, data)
}
\arguments{
\item{SeedN}{Name of the colonn with the seed numbers}

\item{evalName}{Prefix of the evalaution variable}

\item{data}{The name of the data frame containing the data.}
}
\value{
Data frame with the summary values of germination variables.
}
\description{
This function makes a data table with the result of germination variables for each experimental unit.
}
\examples{

library(GerminaR)
dt <- GerminaR
smr <- ger_summary(SeedN = "NSeeds", evalName = "Ev", data = dt)
smr
}

