% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ger_summary.R
\name{ger_summary}
\alias{ger_summary}
\title{Summary of Germination indices}
\usage{
ger_summary(factors, SeedN, evalName, cumulative = FALSE, data)
}
\arguments{
\item{factors}{Factor included for the analysis.}

\item{SeedN}{Name of the column with the seed numbers.}

\item{evalName}{Prefix of the evaluation variable.}

\item{cumulative}{Type of data collection \verb{[logic: FALSE or TRUE]}}

\item{data}{The name of the data frame containing the data.}
}
\value{
Data frame with the summary values of germination variables.
}
\description{
This function makes a data table with the result of germination indices for each experimental unit.
}
\examples{

library(GerminaR)
fb <- prosopis
smr <- ger_summary(factors = c("nacl", "temp", "rep")
                   , SeedN = "seeds"
                   , evalName = "D"
                   , cumulative = FALSE
                   , data = fb)
smr

}
