% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexes.R
\name{ger_VGT}
\alias{ger_VGT}
\title{Variance of the Mean Germination Time}
\usage{
ger_VGT(evalName, data)
}
\arguments{
\item{evalName}{Prefix of the names of the periods of evaluation.}

\item{data}{The name of the data frame containing the data.}
}
\value{
It returns an vector with the values of Variance of Germination
}
\description{
This function calculates the variance of the mean germination time.
}
\examples{
library(GerminaR)
dfr <- prosopis
vgt <- ger_VGT(evalName = "D", data = dfr)
vgt
}
