% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{GerminaQuant}
\alias{GerminaQuant}
\title{@description GermiQuant App allows make the calculation for the germination variables incredibly easy in a interactive applications build in base a GerminaR R package and Shiny.
 GermiQuant App is live!. Outputs change instantly as users modify inputs, without requiring a reload the app.
 The principal features of the application allow calculate the principal germination variables, statistical analysis and easy way to plot the results.}
\usage{
GerminaQuant()
}
\description{
@description GermiQuant App allows make the calculation for the germination variables incredibly easy in a interactive applications build in base a GerminaR R package and Shiny.
 GermiQuant App is live!. Outputs change instantly as users modify inputs, without requiring a reload the app.
 The principal features of the application allow calculate the principal germination variables, statistical analysis and easy way to plot the results.
}
\concept{GerminaR}
