% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{GerminaQuant}
\alias{GerminaQuant}
\title{@description GermiQuant for R app allows make the calculation for the germination indices incredibly easy in a interactive applications build in base a GerminaR R package and Shiny.
 GermiQuant app is live!. Outputs change instantly as users modify inputs, without requiring a reload the app.
 The principal features of the application allow calculate the principal germination indices, statistical analysis and easy way to plot the results.}
\usage{
GerminaQuant()
}
\description{
@description GermiQuant for R app allows make the calculation for the germination indices incredibly easy in a interactive applications build in base a GerminaR R package and Shiny.
 GermiQuant app is live!. Outputs change instantly as users modify inputs, without requiring a reload the app.
 The principal features of the application allow calculate the principal germination indices, statistical analysis and easy way to plot the results.
}
\concept{GerminaR}
