% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germ_datasets.R
\docType{data}
\name{GerminaR}
\alias{GerminaR}
\title{Germination Data from salinity experiment in Jatropha}
\format{A data frame with 100 columns and 100 rows.}
\source{
UFRPE experimental data.
}
\usage{
GerminaR
}
\description{
This data set contains a dataset from 26 days of germination experiment
Jatropha Curcas.
}
\references{
This data set was developed in UFRPE in 2015
}

