% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexes.R
\name{GRP}
\alias{GRP}
\title{Germination Seed percentage}
\usage{
GRP(seeds, values)
}
\arguments{
\item{seeds}{Total seed sown}

\item{values}{Germination values}
}
\value{
Percentage of seed germinated
}
\description{
This function Calculate the Germination percentage
}
\author{
Flavio Lozano
}

