% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexes.R
\name{GRU}
\alias{GRU}
\title{Germination Uncertainty}
\usage{
GRU(values)
}
\arguments{
\item{values}{Germination values}
}
\value{
Germination Uncertainty (U)
}
\description{
This function Calculate the Germination Uncertainty
}
\author{
Flavio Lozano
}

