% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{metamorphosis}
\alias{metamorphosis}
\title{Transform dataframe based in a dictionary}
\usage{
metamorphosis(fieldbook, dictionary, from, to, index, colnames)
}
\arguments{
\item{fieldbook}{data frame with the original information}

\item{dictionary}{data frame with at least 3 colums (original names, new names and variable type)}

\item{from}{Column name of a data frame with the original names of the variables}

\item{to}{Column name of a data frame with the new names for the variables}

\item{index}{Column name of a data frame with the type and level of the variables}

\item{colnames}{Character vector with the column names}
}
\value{
List with original fieldbook, variables and new fieldbook
}
\description{
Transfor all fieldbook data frame according to data dictionary
}
