% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexes.R
\name{MGT}
\alias{MGT}
\title{Mean Germination Time}
\usage{
MGT(eval_days)
}
\arguments{
\item{eval_days}{Germination eval_days}
}
\value{
Mean Germination Time
}
\description{
This function Calculate the Mean Germination Time
}
\author{
Flavio Lozano
}

