% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexes.R
\name{SDG}
\alias{SDG}
\title{Standard deviation of the Germination Time}
\usage{
SDG(eval_days)
}
\arguments{
\item{eval_days}{Germination eval_days}
}
\value{
Standard deviation of the Germination Time
}
\description{
This function Calculate the Standard deviation of the Germination Time
}
\author{
Flavio Lozano
}

