% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexes.R
\name{VGT}
\alias{VGT}
\title{Variance of Germination Time}
\usage{
VGT(values)
}
\arguments{
\item{values}{Germination values}
}
\value{
Variance of Germination Time
}
\description{
This function Calculate the Variance of Germination Time
}
\author{
Flavio Lozano
}

