% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webTable.R
\name{webTable}
\alias{webTable}
\title{Web table}
\usage{
webTable(
  data,
  caption = NULL,
  digits = 3,
  rnames = FALSE,
  buttons = NULL,
  file_name = NULL,
  scrolly = NULL
)
}
\arguments{
\item{data}{Data set.}

\item{caption}{Title for the table.}

\item{digits}{Digits number in the table exported.}

\item{rnames}{Row names.}

\item{buttons}{Buttons \verb{[character: excel, copy, none]}.}

\item{file_name}{Excel file name.}

\item{scrolly}{Windows height to show the table.}
}
\description{
Web table in HTML format
}
